requirements.txt
Django==4.1.1
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11